Refatoração
Funções repetidas modularizadas para compartilhamento quando necessário!
