+++ src/controllers/user/registerController.ts
@@ -1,11 +1,11 @@
 import { Request, Response } from "express";
 import bcrypt from "bcrypt";
 import userModel from "../../models/user-model";
-import { UserData } from "../../interfaces/user-data";
+import { validateUserData } from "../../utils/userValidation";
 
 export const register = async (req: Request, res: Response) => {
   try {
-    const data = await readData(req.body);
+    const data = await validateUserData(req.body);
 
     if (typeof data === "string") {
       return res.status(400).json({ message: data });
@@ -28,33 +28,5 @@ export const register = async (req: Request, res: Response) => {
     res.status(201).json({ message: "Usuário criado com sucesso!" });
   } catch (err) {
     res.status(500).json({ message: `Erro interno do servidor: ${err}` });
-  }
-
-  async function readData(userData: UserData) {
-    const { username, password, email } = userData;
-
-    if (!username || !password || !email) {
-      return "Preencha todos os campos!";
-    }
-
-    if (username !== username.trim() || password !== password.trim()) {
-      return "Os dados não podem começar com espaçamentos";
-    }
-
-    if (username.length < 4 || username.length > 16) {
-      return "O nome de usuário deve ter entre 4 e 16 caracteres!";
-    }
-
-    const emailRegex = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
-
-    if (!emailRegex.test(email)) {
-      return "Insira um email válido!";
-    }
-
-    if (password.length < 6 || password.length > 20) {
-      return "A senha deve ter entre 6 e 20 caracteres!";
-    }
-
-    return userData;
   }
 };
