+++ src/controllers/user/updateController.ts
@@ -1,8 +1,7 @@
 import { Request, Response } from "express";
 import mongoose from "mongoose";
-import bcrypt from "bcrypt";
 import userModel from "../../models/user-model";
-import { UserData } from "../../interfaces/user-data";
+import { updateData, validateUserData } from "../../utils/userValidation";
 
 export const update = async (req: Request, res: Response) => {
   try {
@@ -21,7 +20,7 @@ export const update = async (req: Request, res: Response) => {
       return res.status(404).json({ message: "Usuário não encontrado!" });
     }
 
-    const data = await readData(req.body);
+    const data = await validateUserData(req.body);
 
     if (typeof data === "string") {
       return res.status(400).json({ message: data });
@@ -31,52 +30,5 @@ export const update = async (req: Request, res: Response) => {
     res.status(200).json({ message: "Save atualizado com sucesso!" });
   } catch (err) {
     res.status(500).json({ message: `Erro interno no servidor: ${err}` });
-  }
-
-  async function readData(userData: UserData) {
-    const { username, password, email } = userData;
-
-    if (!username && !password && !email) {
-      return "Nenhum dado fornecido para atualização!";
-    }
-
-    if (
-      (username && username !== username.trim()) ||
-      (password && password !== password.trim())
-    ) {
-      return "Os dados não podem começar com espaçamentos";
-    }
-
-    if (username && (username.length < 4 || username.length > 16)) {
-      return "O nome de usuário deve ter entre 4 e 16 caracteres!";
-    }
-
-    if (email) {
-      const emailRegex = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
-
-      if (!emailRegex.test(email)) {
-        return "Insira um email válido!";
-      }
-    }
-
-    if (password && (password.length < 6 || password.length > 20)) {
-      return "A senha deve ter entre 6 e 20 caracteres!";
-    }
-
-    return userData;
-  }
-
-  async function updateData(userData: UserData, id: mongoose.Types.ObjectId) {
-    let updateData: Partial<UserData> = {};
-
-    if (userData.username) updateData.username = userData.username;
-    if (userData.email) updateData.email = userData.email.toLowerCase();
-
-    if (userData.password) {
-      const newPassword = await bcrypt.hash(userData.password, 10);
-      updateData.password = newPassword;
-    }
-
-    await userModel.findByIdAndUpdate(id, { $set: updateData });
   }
 };
